feat: domian -> domain 명명
