+++ src/main/java/org/running/domain/board/RunningApplication.java
@@ -1,0 +1,13 @@
+package org.running.domain.board;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RunningApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(RunningApplication.class, args);
+    }
+
+}
