+++ src/main/java/org/running/domain/board/model/DTO/request/BoardDeleteRequest.java
@@ -1,0 +1,8 @@
+package org.running.domain.board.model.DTO.request;
+
+import lombok.Data;
+
+@Data
+public class BoardDeleteRequest {
+    private Long boardNumber;
+}
