+++ src/main/java/org/running/domain/board/model/DTO/request/ReplyDeleteRequest.java
@@ -1,0 +1,10 @@
+package org.running.domain.board.model.DTO.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ReplyDeleteRequest {
+    private Long replyNumber;
+}
