+++ src/main/java/org/running/domain/board/model/DTO/request/ReplyPostRequest.java
@@ -1,0 +1,9 @@
+package org.running.domain.board.model.DTO.request;
+
+import lombok.Data;
+
+@Data
+public class ReplyPostRequest {
+    private Long boardNumber;
+    private String content;
+}
