+++ src/main/java/org/running/domain/board/model/DTO/response/BoardListResponse.java
@@ -1,0 +1,19 @@
+package org.running.domain.board.model.DTO.response;
+
+import org.running.domain.board.model.entity.Board;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class BoardListResponse {
+    private Long boardNumber;
+    private String title;
+
+    public static BoardListResponse from(Board board){
+        return new BoardListResponse(
+                board.getBoardNumber(),
+                board.getTitle()
+        );
+    }
+}
