+++ src/main/java/org/running/domain/board/model/DeleteStatus.java
@@ -1,0 +1,6 @@
+package org.running.domain.board.model;
+
+public enum DeleteStatus {
+
+    ACTIVE,DELETE;
+}
