+++ src/main/java/org/running/domain/user/controller/UserApiController.java
@@ -1,0 +1,1 @@
+package org.running.domain.user.controller;
