+++ src/main/java/org/running/domain/user/dto/AddUserRequest.java
@@ -1,0 +1,11 @@
+package org.running.domain.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AddUserRequest {
+    private String email;
+    private String password;
+}
