--- src/main/java/org/running/domian/board/RunningApplication.java
@@ -1,13 +1,0 @@
-package org.running.domian.board;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RunningApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(RunningApplication.class, args);
-    }
-
-}
