--- src/main/java/org/running/domian/board/model/DTO/request/BoardDeleteRequest.java
@@ -1,8 +1,0 @@
-package org.running.domian.board.model.DTO.request;
-
-import lombok.Data;
-
-@Data
-public class BoardDeleteRequest {
-    private Long boardNumber;
-}
