--- src/main/java/org/running/domian/board/model/DTO/request/BoardPostRequest.java
@@ -1,12 +1,0 @@
-package org.running.domian.board.model.DTO.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class BoardPostRequest {
-    private String title;
-    private String content;
-
-}
