--- src/main/java/org/running/domian/board/model/DTO/request/ReplyDeleteRequest.java
@@ -1,10 +1,0 @@
-package org.running.domian.board.model.DTO.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ReplyDeleteRequest {
-    private Long replyNumber;
-}
