--- src/main/java/org/running/domian/board/model/DTO/request/ReplyModifyRequest.java
@@ -1,12 +1,0 @@
-package org.running.domian.board.model.DTO.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ReplyModifyRequest {
-
-    private Long replyNumber;
-    private String content;
-}
