--- src/main/java/org/running/domian/board/model/DTO/request/ReplyPostRequest.java
@@ -1,9 +1,0 @@
-package org.running.domian.board.model.DTO.request;
-
-import lombok.Data;
-
-@Data
-public class ReplyPostRequest {
-    private Long boardNumber;
-    private String content;
-}
