--- src/main/java/org/running/domian/board/model/DeleteStatus.java
@@ -1,6 +1,0 @@
-package org.running.domian.board.model;
-
-public enum DeleteStatus {
-
-    ACTIVE,DELETE;
-}
