--- src/main/java/org/running/domian/user/dto/AddUserRequest.java
@@ -1,6 +1,0 @@
-package org.running.domian.user.dto;
-
-public class AddUserRequest {
-    private String email;
-    private String password;
-}
