--- src/main/java/org/running/domian/user/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package org.running.domian.user.repository;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email); // email로 사용자 정보 가져옴
-}
