--- src/main/java/org/running/domian/user/service/UserService.java
@@ -1,1 +1,0 @@
-package org.running.domian.user.service;
